pyproject.toml
[project]
name = "audiocoqui"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "pydub>=0.25.1",
    "pypdf2>=3.0.1",
    "python-dotenv>=1.0.1",
    "rich-click>=1.8.5",
    "rich>=13.9.4",
    "tqdm>=4.67.1",
    "tts>=0.22.0",
]
